Add check for RecognitionSuccessType.UNSUCCESSFUL
This fixes a situation, when empty "successful" result is being
reported, when it really should not be reported.

For example, when scanning a passport in a way that two pages of it
are scanned, we were getting "Turn over" notification and empty
result. This resulted in a user being stuck. Now, no "turn over"
notification will be reported (as no result will be reported), and
scan will only succeed when only one page of a passport is being
scanned.