fix: [RND-175929] Fix unnecessary scan requests on Android (#12)
We use our own camera to give the images to Microblink scanner. The
images are produced on every camera frame. When we receive the image,
we first check whether Microblink document scanner is ready. If it is
not ready, we basically discard the image. If Microblink scanner is
ready to accept the images, we convert the image that camera gives us
to bitmap and then we give this bitmap to Microblink scanner to
scan. With this bitmap we also set a listener that has two callbacks -
to receive the results of the scan. The first callback is for when we
hit unrecoverable error. In this case, we signal to Flutter side that
we hit the unrecoverable error and we reset the recognition state of
the scanner (we do not close the image, hence we are not getting a new
one to scan, because Flutter should close the scanner view on
unrecoverable error). The other callback is onScanningDone. Basically,
this callback is called when Microblink has some result. The result
has recognition success type. There are only four recognition success
types. It is unsuccessful, partial, successful and
stage_successful. We check that the recognition success type is not
unsuccessful, and signal to Flutter side that we have some result for
it. Then we start accepting the images from camera again. And this is
the root cause of the issue: since Flutter accepts the result of the
scan asynchronously and we immediately start scanning new images, we
might get more scans than we need before the Android View is closed by
Flutter.

Solution: do not close the image (image.close()) immediately after
sending the result to Flutter. Instead, wait for when Flutter responds
that it processed the result and is awaiting more scans. When getting
this notification from Flutter, close the image. That would result in
receiving more images from camera for analysis and scanning.